backend/agent/juris/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud dependencies
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1

# LangChain dependencies
langchain==0.1.0
langchain-google-vertexai==0.1.0
langchain-community==0.0.10

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
unstructured==0.11.8

# Additional utilities
requests==2.31.0
numpy==1.24.3
pandas==2.1.4
